+++ CHANGELOG.md
@@ -47,6 +47,7 @@ This project adheres to [Semantic Versioning](https://semver.org).
 	* Add ``_parse_frame_data`` helper method to all ID3v2 frame classes.
 	* ``ID3v2Frame.parse`` calls into helper methods for appropriate subclass.
 * Revise exceptions.
+* Check for framing bit in Ogg Vorbis comments.
 
 ### Removed